ensure valid header config in `AuthConfig` (#1656)
<!-- The PR description should answer 2 important questions: -->

### What

This PR adds a validation on the auth config headers config to ensure:
- Valid header name
- Valid header value
This PR also fixes the bug where the auth hook headers filtering was
case-sensitive.
### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

We have added a new validate function and have used a compatibility flag
to ensure backwards compatibility.

Previously, we were going over the header list and checked (implicitly
converting the header to a lowercase string) wheather the header was in
the forward list or not.
Now, we are going over the forward list, creating a case-insensitive
header name and then checking if the header exists in the client
headers.

This ensures that the header lookup is case-insensitive.

---------

Co-authored-by: Daniel Chambers <daniel@hasura.io>
V3_GIT_ORIGIN_REV_ID: 9bb4c013a5bfb12c927b020a0e2f411bb417a0cd