Fix remote joins involving mapped fields that contain object types (#1724)
### What
This PR fixes two related issues when performing a remote join that
contains a mapping between fields that contain an object type.

1. If the field being joined upon is selected in the query, then only
what is selected in the query is sent to the target data connector
2. If the target data connector's object type has different field names
than the source data connector, the join fails because the join simply
sends the object with the source's field names, not the target's field
names.

Joining on a field that contains an object type is more common when
"joining" to command backed by a NodeJS Lambda function, where the
purpose is to pass a whole object of stuff to the function for it to do
something with.

### How

The first issue is caused by an optimisation that causes remote joins to
reuse the existing field selection if it already exists in the query. If
the user has not selected the entire object, then this subset is what is
reused, which is incorrect. The solution is to detect if the user has
performed a subselection on the nested object, and if so, don't reuse
the field and instead introduce a phantom field to select the whole
nested object unmolested.

The second issue is fixed by capturing mappings from source object type
field names to target object type field names during planning and then
using those mapping during remote join execution to map the source
object to the target shape.

A new test `test_remote_relationships_command_join_with_object_value`
checks this functionality by joining from one custom connector to
another. It takes a location from an institution in the first connector
and passes it to a new "eval_location" function in the second connector.
The wrinkle is that "eval_location" takes an object type with different
names, so the source `location` type has to be mapped correctly to the
target `location_pascalcase` type.

V3_GIT_ORIGIN_REV_ID: 3b2c88930c2a213a8b03bcbdcdcdd4de00bc51d7