Improve MBS error if CLI mangles the DataConnectorLink (#1732)
### What

Older versions of the ddn CLI do not understand NDC 0.2.0, and so when
they introspect a connector to update the DataConnectorLink, they
blindly deserialize NDC 0.2.0 schema and capabilities into 0.1.x types,
which mangles them. They then write the mangled result into the
DataConnectorLink. If the user then performs a build, they get an
unhelpful JSON deserialization error complaining about the mangled
DataConnectorLink.

This PR detects this situation (a deserialization error coming from the
schema section of a DataConnectorLink) and replaces the standard error
message with a more helpful one that guides the user to upgrade their
ddn CLI version.

![image](https://github.com/user-attachments/assets/9f50e8a0-fb4e-468b-a4b2-1ff7c9f75a18)

### How

During OpenDD object deserialization (known as "refinement"), we detect
if the object being deserialized is a DataConnectorLink. Then, if
there's a deserialization error, we check if the problem was in the
`definition.schema` part of the object. If so, we update the error
message with a more helpful one.

V3_GIT_ORIGIN_REV_ID: a0662b2e57ea5d60880fc42d8eb6527d3bcf347e