Pushdown via plan (#1753)
<!-- The PR description should answer 2 important questions: -->

### What

This adds an experimental pushdown path which generates a `Plan` object,
which is sent to an undocumented `/plan` endpoint on the connector. It
is enabled for now via a `SQL_CONNECTORS` env var, which we plan to
remove. We also plan to implement this all properly as a bona fide
feature in the NDC spec, but that will come later.

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->

<!-- Consider: do we need to add a changelog entry? -->

<!-- Does this PR introduce new validation that might break old builds?
-->

<!-- Consider: do we need to put new checks behind a flag? -->

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

---------

Co-authored-by: Vamshi Surabhi <0x777@users.noreply.github.com>
V3_GIT_ORIGIN_REV_ID: 46f85b48cf2e2ad0bde483ece59f15253215aff5