Store operator meanings in resolved model predicate (#1816)
<!-- The PR description should answer 2 important questions: -->

### What

We need to understand the meaning of each comparison operator to convert
`ModelPredicate` into something `sql` understands. It's easier to
calculate them when we have all the information available and store it
in the `ModelPredicate` than reconstruct it on every request, so this
adds it to `metadata-resolve`.

V3_GIT_ORIGIN_REV_ID: 67c4822cf7075cab417ef491c3e86eeaf3fa2ffe