More helpful comparable relationships error (#1836)
<!-- The PR description should answer 2 important questions: -->

### What

When a `comparableRelationship` on a boolean expression type is defined,
but the target model did not have a `filterExpressionType` defined, we
acted as if the `comparableRelationship` was not defined, which is
really confusing.

This PR
a) adds a warning when such a thing that is set up, that turns into an
error when the compat flag is on
b) makes the `boolean_expression_type` in a resolved comparable
relationship optional, so we can tell the difference between "it's
defined but there's nothing there" and "it's not defined"

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

V3_GIT_ORIGIN_REV_ID: 592385b9c942a39ff838130bb72d2e41790095d8