Enable remote predicates in SQL layer (#1889)
<!-- The PR description should answer 2 important questions: -->

### What

Previously the SQL was silently dropping remote predicates, meaning a
select permission that uses them would not work. We did not advertise
this feature, as such, but there was nothing stopping a user from
setting one up.

Stacked on #1890 because the tests here rely on a `metadata-resolve`
fix.

This PR:
- Makes the SQL layer execute any remote predicates that are found. This
functionality is already used and tested extensively in the GraphQL
layer, so it feels safe to do so.
- Makes sure that relational query pushdown fails when a permission
using a remote relationship is found

A note on remote joins in SQL:
- Actual remote joins are not used in SQL (we never construct
`open_dds::query::RelationshipSelection` or
`open_dds::query::RelationshipAggregateSelection` IR), but if they
existed we would execute them now.

V3_GIT_ORIGIN_REV_ID: 9d76586a91be28807e733a734b8123ed0be65dda