Include NDC scalar representation in argument info (#1910)
<!-- The PR description should answer 2 important questions: -->

### What

Include the NDC scalar representation of an argument in `ArgumentInfo`
in `metadata-resolve`.

### How

This requires a bit of moving stuff around. We already had an
`arguments` stage which knows about the data connector each model /
command uses, so we resolve this information there and add it to
`ModelWithPermission` and `CommandWithPermission` lifted outside the
`Model` and `Command` types.

As a nice side effect we remove a bit of unused stuff from the final
resolved `Model` type.

V3_GIT_ORIGIN_REV_ID: 876a46fc3caa777b7473fc306a79a1879e7bb239