normalize ndc client usage (#1966)
In preparation for the the ndc plugins, this PR normalizes the public
interface of the execute crate.

Specifically, we make the functions in client.rs `pub (crate)` so they
can't be used directly. We then add two functions for explain requests,
similar to those used for query and mutation requests.

We also bind the result of the trace wrapper to a variable and return
`Ok(response)` in prepartion to add the plugins in between.

functionally, this PR should not change anything

V3_GIT_ORIGIN_REV_ID: a8780eb488ad69b337ca1429337f7fb2b1ea1adb