Implement ndc plugins (#1978)
<!-- The PR description should answer 2 important questions: -->

### What

<!-- What is this PR trying to accomplish (and why, if it's not
obvious)? -->

This PR implements ndc plugins, both pre-request and pre-response.

<!-- Consider: do we need to add a changelog entry? -->

<!-- Does this PR introduce new validation that might break old builds?
-->

<!-- Consider: do we need to put new checks behind a flag? -->

### How

<!-- How is it trying to accomplish it (what are the implementation
steps)? -->

In the execute crate, we add plugin execution in the functions that
reach out to the connectors.
We cover the endpoints `/query`, `/query/explain`, `/mutation`,
`/mutation/explain`

Note we do not cover any other endpoints.
Also note the relevant metadata is currently hidden. I'm not sure if we
want to change that now, or later in a separate PR for the actual
release. Same for changelog and docs.

V3_GIT_ORIGIN_REV_ID: dde98d50916396d14f366449c67f462fa8956a01