Make 'Built-in operators require a boolean expression type' a user error (#2098)
<!-- The PR description should answer 2 important questions: -->

### What

`Built-in operators require a boolean expression type` shows up as an
internal error, when it's really because a boolean expression type
hasn't been provided for a type. Change the type to reflect this, and
include the object type name in question to make it easier to fix.

V3_GIT_ORIGIN_REV_ID: 89fae7521fd9fe1135bfb5c8c9d44c07fc1d93cb