Fix `and` nested in `exists` for filtering (#2135)
This fixes an issue in engine where the following filters were treated
equivalently:

```graphql
users: { _and: [
  { first_name: { _eq: "Bruce" },
  { last_name: { _eq: "Willis" }
]}
```

```graphql
_and: [
  { users: { first_name: { _eq: "Bruce" } },
  { users: { last_name: { _eq: "Willis" } }
]}
```

The first means "return something where there is a `user` named 'Bruce
Willis'", the second means "return something where there is a `user`
with first name 'Bruce' and a `user` with last name 'Willis'".

We always treated both as the second case, now we treat them correctly.

This was because the OpenDD IR for `BooleanExpression` did not allow
nesting inside a nested array, we could not express the idea that the
`AND` is inside the `EXISTS`. This is now fixed in GraphQL. Filtering
across nested arrays is not expressible in SQL frontend so not a concern
there.

V3_GIT_ORIGIN_REV_ID: 2f26b165ae7e39dd5c41e6e60a2e2c32ed70115c